// frontend/src/api/generated/model/tutor-dto.ts
/**
 * EduMatch API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TutorSubjectDto } from './tutor-subject-dto';
import { AddressDto } from './address-dto';
import { Gender } from './gender';


export interface TutorDto { 
    id?: number;
    code?: string | null;
    userId?: number;
    fullName?: string | null;
    avatarUrl?: string | null;
    gender?: Gender;
    hourlyRate?: number;
    rating?: number;
    totalReviews?: number;
    address?: AddressDto;
    subjects?: Array<TutorSubjectDto> | null;
}